scripts/requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
fastjsonschema==2.21.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
langdetect==1.0.9
matplotlib-inline==0.1.7
narwhals==1.34.1
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
spotipy==2.25.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
